User Input Validation
This project created in pure javscript with Html and css. When you load the page, It has an user name and password and an submit button. If you have entered the correct username and password, it prints a message as "Your username and password is correct" message below the HTML page. If you entered your details are incorrect, It will shown the following message, below the button,  "You have entered wrong username and password".